preprocess/pyproject.toml
[project]
name = "preprocess"
version = "0.1.0"
description = ""
authors = [
    {name = "Zuhanit",email = "zuhanit3@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy (>=2.2.4,<3.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "pydantic (>=2.11.1,<3.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "rectpack (>=0.2.2,<0.3.0)",
    "rich (>=14.0.0,<15.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
